GLIDE-4 Remaining from upstream merge. refactored routes and bind
